+++ python/boomer/common/cpp/predictions.h
@@ -17,13 +17,13 @@ class Prediction {
 
         /**
          * @param numPredictions    The number of labels for which the rule predicts
-         * @param labelIndices      A pointer to an array of type `intp`, shape `(numPredictions)`, representing the
+         * @param labelIndices      A pointer to an array of type `uint32`, shape `(numPredictions)`, representing the
          *                          indices of the labels for which the rule predicts of NULL, if the rule predicts for
          *                          all labels
          * @param predictedScores   A pointer to an array of type `float64`, shape `(numPredictions)`, representing the
          *                          predicted scores
          */
-        Prediction(intp numPredictions, intp* labelIndices, float64* predictedScores);
+        Prediction(intp numPredictions, uint32* labelIndices, float64* predictedScores);
 
         ~Prediction();
 
@@ -33,10 +33,10 @@ class Prediction {
         intp numPredictions_;
 
         /**
-         * A pointer to an array of type `intp`, shape `(numPredictions_)`, representing the indices of the labels for
+         * A pointer to an array of type `uint32`, shape `(numPredictions_)`, representing the indices of the labels for
          * which the rule predicts or NULL, if the rule predicts for all labels.
          */
-        intp* labelIndices_;
+        uint32* labelIndices_;
 
         /**
          * A pointer to an array of type `float64`, shape `(numPredictions_)`, representing the predicted scores.
@@ -54,14 +54,14 @@ class PredictionCandidate : public Prediction {
 
         /**
          * @param numPredictions        The number of labels for which the rule predicts
-         * @param labelIndices          A pointer to an array of type `intp`, shape `(numPredictions)`, representing the
-         *                              indices of the labels for which the rule predicts of NULL, if the rule predicts
-         *                              for all labels
+         * @param labelIndices          A pointer to an array of type `uint32`, shape `(numPredictions)`, representing
+         *                              the indices of the labels for which the rule predicts of NULL, if the rule
+         *                              predicts for all labels
          * @param predictedScores       A pointer to an array of type `float64`, shape `(numPredictions)`, representing
          *                              the predicted scores
          * @param overallQualityScore   A score that assesses the overall quality of the predictions
          */
-        PredictionCandidate(intp numPredictions, intp* labelIndices, float64* predictedScores,
+        PredictionCandidate(intp numPredictions, uint32* labelIndices, float64* predictedScores,
                             float64 overallQualityScore);
 
         ~PredictionCandidate();
@@ -82,16 +82,16 @@ class LabelWisePredictionCandidate : public PredictionCandidate {
 
         /**
          * @param numPredictions        The number of labels for which the rule predicts
-         * @param labelIndices          A pointer to an array of type `intp`, shape `(numPredictions)`, representing the
-         *                              indices of the labels for which the rule predicts of NULL, if the rule predicts
-         *                              for all labels
+         * @param labelIndices          A pointer to an array of type `uint32`, shape `(numPredictions)`, representing
+         *                              the indices of the labels for which the rule predicts of NULL, if the rule
+         *                              predicts for all labels
          * @param predictedScores       A pointer to an array of type `float64`, shape `(numPredictions)`, representing
          *                              the predicted scores
          * @param qualityScores         A pointer to an array of type `float64`, shape `(numPredictions)`, representing
          *                              the quality scores for individual labels
          * @param overallQualityScore   A score that assesses the overall quality of the predictions
          */
-        LabelWisePredictionCandidate(intp numPredictions, intp* labelIndices, float64* predictedScores,
+        LabelWisePredictionCandidate(intp numPredictions, uint32* labelIndices, float64* predictedScores,
                                      float64* qualityScores, float64 overallQualityScore);
 
         ~LabelWisePredictionCandidate();
