+++ python/boomer/seco/cpp/label_wise_statistics.h
@@ -28,7 +28,7 @@ namespace seco {
 
             intp numPredictions_;
 
-            const intp* labelIndices_;
+            const uint32* labelIndices_;
 
             std::shared_ptr<AbstractRandomAccessLabelMatrix> labelMatrixPtr_;
 
@@ -53,7 +53,7 @@ namespace seco {
              *                                  to be used for calculating the predictions, as well as corresponding
              *                                  quality scores, of rules
              * @param numPredictions            The number of labels to be considered by the search
-             * @param labelIndices              An array of type `intp`, shape `(numPredictions)`, representing the
+             * @param labelIndices              An array of type `uint32`, shape `(numPredictions)`, representing the
              *                                  indices of the labels that should be considered by the search or NULL,
              *                                  if all labels should be considered
              * @param labelMatrix               A shared pointer to an object of type `AbstractRandomAccessLabelMatrix`
@@ -73,7 +73,7 @@ namespace seco {
              *                                  the previous refinement of the rule, for each label
              */
             DenseLabelWiseRefinementSearchImpl(std::shared_ptr<AbstractLabelWiseRuleEvaluation> ruleEvaluationPtr,
-                                               intp numPredictions, const intp* labelIndices,
+                                               intp numPredictions, const uint32* labelIndices,
                                                std::shared_ptr<AbstractRandomAccessLabelMatrix> labelMatrixPtr,
                                                const float64* uncoveredLabels, const uint8* minorityLabels,
                                                const float64* confusionMatricesTotal,
@@ -166,7 +166,7 @@ namespace seco {
 
             void updateCoveredStatistic(intp statisticIndex, uint32 weight, bool remove) override;
 
-            AbstractRefinementSearch* beginSearch(intp numLabelIndices, const intp* labelIndices) override;
+            AbstractRefinementSearch* beginSearch(intp numLabelIndices, const uint32* labelIndices) override;
 
             void applyPrediction(intp statisticIndex, Prediction* prediction) override;
